mtx: move out coinselector.
