+++ lib/primitives/mtx.js
@@ -8,7 +8,6 @@
 
 const assert = require('bsert');
 const {encoding} = require('bufio');
-const {BufferMap} = require('buffer-map');
 const Script = require('../script/script');
 const TX = require('./tx');
 const Input = require('./input');
@@ -18,14 +17,17 @@ const Outpoint = require('./outpoint');
 const CoinView = require('../coins/coinview');
 const Path = require('../wallet/path');
 const WalletCoinView = require('../wallet/walletcoinview');
-const Address = require('./address');
 const consensus = require('../protocol/consensus');
 const policy = require('../protocol/policy');
-const Amount = require('../ui/amount');
 const Stack = require('../script/stack');
 const rules = require('../covenants/rules');
 const util = require('../utils/util');
 const {types} = rules;
+
+const {
+  CoinSelector,
+  InMemoryCoinSource
+} = require('../utils/coinselector');
 
 /** @typedef {import('../types').SighashType} SighashType */
 /** @typedef {import('../types').Hash} Hash */
@@ -34,6 +36,8 @@ const {types} = rules;
 /** @typedef {import('../protocol/network')} Network */
 /** @typedef {import('../workers/workerpool')} WorkerPool */
 /** @typedef {import('./keyring')} KeyRing */
+/** @typedef {import('./address')} Address */
+/** @typedef {import('../utils/coinselector')} coinselector */
 
 /**
  * MTX
@@ -1018,14 +1022,26 @@ class MTX extends TX {
   /**
    * Select necessary coins based on total output value.
    * @param {Coin[]} coins
-   * @param {Object?} options
+   * @param {Object} options
    * @returns {Promise<CoinSelector>}
    * @throws on not enough funds available.
    */
 
-  selectCoins(coins, options) {
-    const selector = new CoinSelector(this, options);
-    return selector.select(coins);
+  async selectCoins(coins, options) {
+    const source = new InMemoryCoinSource({
+      coins,
+      selection: options.selection
+    });
+
+    await source.init();
+
+    if (options.selection === 'all')
+      options.selectAll = true;
+
+    const selector = new CoinSelector(this, source, options);
+    await selector.select();
+
+    return selector;
   }
 
   /**
@@ -1112,7 +1128,9 @@ class MTX extends TX {
   /**
    * Select coins and fill the inputs.
    * @param {Coin[]} coins
-   * @param {Object} options - See {@link MTX#selectCoins} options.
+   * @param {Object} options - See
+   *    {@link coinselector.CoinSelectorOptions} and
+   *    {@link coinselector.CoinSourceOptions} options.
    * @returns {Promise<CoinSelector>}
    */
 
@@ -1175,7 +1193,14 @@ class MTX extends TX {
     const inputs = [];
     /** @type {Output[]} */
     const outputs = [];
-    // [Input, Output][]
+
+    /**
+     * @typedef {Array} Linked
+     * @property {Input} 0
+     * @property {Output} 1
+     */
+
+    /** @type {Linked[]} */
     const linked = [];
 
     let i = 0;
@@ -1411,524 +1436,25 @@ class MTX extends TX {
   }
 }
 
-/**
- * Coin Selector
- * @alias module:primitives.CoinSelector
- */
-
-class CoinSelector {
-  /**
-   * Create a coin selector.
-   * @constructor
-   * @param {MTX} tx
-   * @param {Object?} options
-   */
-
-  constructor(tx, options) {
-    this.tx = tx.clone();
-    this.view = tx.view;
-    this.coins = [];
-    this.outputValue = 0;
-    this.index = 0;
-    this.chosen = [];
-    this.change = 0;
-    this.fee = CoinSelector.MIN_FEE;
-
-    this.selection = 'value';
-    this.subtractFee = false;
-    this.subtractIndex = -1;
-    this.height = -1;
-    this.depth = -1;
-    this.hardFee = -1;
-    this.rate = CoinSelector.FEE_RATE;
-    this.maxFee = -1;
-    this.round = false;
-    this.coinbaseMaturity = 400;
-    this.changeAddress = null;
-    this.inputs = new BufferMap();
-
-    // Needed for size estimation.
-    this.estimate = null;
-
-    this.injectInputs();
-
-    if (options)
-      this.fromOptions(options);
-  }
-
-  /**
-   * Initialize selector options.
-   * @param {Object} options
-   * @private
-   */
-
-  fromOptions(options) {
-    if (options.selection) {
-      assert(typeof options.selection === 'string');
-      this.selection = options.selection;
-    }
-
-    if (options.subtractFee != null) {
-      if (typeof options.subtractFee === 'number') {
-        assert(Number.isSafeInteger(options.subtractFee));
-        assert(options.subtractFee >= -1);
-        this.subtractIndex = options.subtractFee;
-        this.subtractFee = this.subtractIndex !== -1;
-      } else {
-        assert(typeof options.subtractFee === 'boolean');
-        this.subtractFee = options.subtractFee;
-      }
-    }
-
-    if (options.subtractIndex != null) {
-      assert(Number.isSafeInteger(options.subtractIndex));
-      assert(options.subtractIndex >= -1);
-      this.subtractIndex = options.subtractIndex;
-      this.subtractFee = this.subtractIndex !== -1;
-    }
-
-    if (options.height != null) {
-      assert(Number.isSafeInteger(options.height));
-      assert(options.height >= -1);
-      this.height = options.height;
-    }
-
-    if (options.confirmations != null) {
-      assert(Number.isSafeInteger(options.confirmations));
-      assert(options.confirmations >= -1);
-      this.depth = options.confirmations;
-    }
-
-    if (options.depth != null) {
-      assert(Number.isSafeInteger(options.depth));
-      assert(options.depth >= -1);
-      this.depth = options.depth;
-    }
-
-    if (options.hardFee != null) {
-      assert(Number.isSafeInteger(options.hardFee));
-      assert(options.hardFee >= -1);
-      this.hardFee = options.hardFee;
-    }
-
-    if (options.rate != null) {
-      assert(Number.isSafeInteger(options.rate));
-      assert(options.rate >= 0);
-      this.rate = options.rate;
-    }
-
-    if (options.maxFee != null) {
-      assert(Number.isSafeInteger(options.maxFee));
-      assert(options.maxFee >= -1);
-      this.maxFee = options.maxFee;
-    }
-
-    if (options.round != null) {
-      assert(typeof options.round === 'boolean');
-      this.round = options.round;
-    }
-
-    if (options.coinbaseMaturity != null) {
-      assert((options.coinbaseMaturity >>> 0) === options.coinbaseMaturity);
-      this.coinbaseMaturity = options.coinbaseMaturity;
-    }
-
-    if (options.changeAddress) {
-      const addr = options.changeAddress;
-      if (typeof addr === 'string') {
-        this.changeAddress = Address.fromString(addr);
-      } else {
-        assert(addr instanceof Address);
-        this.changeAddress = addr;
-      }
-    }
-
-    if (options.estimate) {
-      assert(typeof options.estimate === 'function');
-      this.estimate = options.estimate;
-    }
-
-    if (options.inputs) {
-      assert(Array.isArray(options.inputs));
-
-      const lastIndex = this.inputs.size;
-      for (let i = 0; i < options.inputs.length; i++) {
-        const prevout = options.inputs[i];
-        assert(prevout && typeof prevout === 'object');
-        const {hash, index} = prevout;
-        this.inputs.set(Outpoint.toKey(hash, index), lastIndex + i);
-      }
-    }
-
-    return this;
-  }
-
-  /**
-   * Attempt to inject existing inputs.
-   * @private
-   */
-
-  injectInputs() {
-    if (this.tx.inputs.length > 0) {
-      for (let i = 0; i < this.tx.inputs.length; i++) {
-        const {prevout} = this.tx.inputs[i];
-        this.inputs.set(prevout.toKey(), i);
-      }
-    }
-  }
-
-  /**
-   * Initialize the selector with coins to select from.
-   * @param {Coin[]} coins
-   */
-
-  init(coins) {
-    this.coins = coins.slice();
-    this.outputValue = this.tx.getOutputValue();
-    this.index = 0;
-    this.chosen = [];
-    this.change = 0;
-    this.fee = CoinSelector.MIN_FEE;
-    this.tx.inputs.length = 0;
-
-    switch (this.selection) {
-      case 'all':
-      case 'random':
-        this.coins.sort(sortRandom);
-        break;
-      case 'age':
-        this.coins.sort(sortAge);
-        break;
-      case 'value':
-        this.coins.sort(sortValue);
-        break;
-      default:
-        throw new FundingError(`Bad selection type: ${this.selection}.`);
-    }
-  }
-
-  /**
-   * Calculate total value required.
-   * @returns {AmountValue}
-   */
-
-  total() {
-    if (this.subtractFee)
-      return this.outputValue;
-    return this.outputValue + this.fee;
-  }
-
-  /**
-   * Test whether the selector has
-   * completely funded the transaction.
-   * @returns {Boolean}
-   */
-
-  isFull() {
-    return this.tx.getInputValue() >= this.total();
-  }
-
-  /**
-   * Test whether a coin is spendable
-   * with regards to the options.
-   * @param {Coin} coin
-   * @returns {Boolean}
-   */
-
-  isSpendable(coin) {
-    if (this.tx.view.hasEntry(coin))
-      return false;
-
-    if (coin.covenant.isNonspendable())
-      return false;
-
-    if (this.height === -1)
-      return true;
-
-    if (coin.coinbase) {
-      if (coin.height === -1)
-        return false;
-
-      if (this.height + 1 < coin.height + this.coinbaseMaturity)
-        return false;
-
-      return true;
-    }
-
-    if (this.depth === -1)
-      return true;
-
-    const depth = coin.getDepth(this.height);
-
-    if (depth < this.depth)
-      return false;
-
-    return true;
-  }
-
-  /**
-   * Get the current fee based on a size.
-   * @param {Number} size
-   * @returns {AmountValue}
-   */
-
-  getFee(size) {
-    // This is mostly here for testing.
-    // i.e. A fee rounded to the nearest
-    // kb is easier to predict ahead of time.
-    if (this.round)
-      return policy.getRoundFee(size, this.rate);
-
-    return policy.getMinFee(size, this.rate);
-  }
-
-  /**
-   * Fund the transaction with more
-   * coins if the `output value + fee`
-   * total was updated.
-   */
-
-  fund() {
-    // Ensure all preferred inputs first.
-    this.resolveInputCoins();
-
-    if (this.isFull())
-      return;
-
-    while (this.index < this.coins.length) {
-      const coin = this.coins[this.index++];
-
-      if (!this.isSpendable(coin))
-        continue;
-
-      this.tx.addCoin(coin);
-      this.chosen.push(coin);
-
-      if (this.selection === 'all')
-        continue;
-
-      if (this.isFull())
-        break;
-    }
-  }
-
-  /**
-   * Initiate selection from `coins`.
-   * @param {Coin[]} coins
-   * @returns {Promise<CoinSelector>}
-   */
-
-  async select(coins) {
-    this.init(coins);
-
-    if (this.hardFee !== -1) {
-      this.selectHard();
-    } else {
-      // This is potentially asynchronous:
-      // it may invoke the size estimator
-      // required for redeem scripts (we
-      // may be calling out to a wallet
-      // or something similar).
-      await this.selectEstimate();
-    }
-
-    if (!this.isFull()) {
-      // Still failing to get enough funds.
-      throw new FundingError(
-        'Not enough funds.',
-        this.tx.getInputValue(),
-        this.total());
-    }
-
-    // How much money is left after filling outputs.
-    this.change = this.tx.getInputValue() - this.total();
-
-    return this;
-  }
-
-  /**
-   * Initialize selection based on size estimate.
-   */
-
-  async selectEstimate() {
-    // Set minimum fee and do
-    // an initial round of funding.
-    this.fee = CoinSelector.MIN_FEE;
-    this.fund();
-
-    // Add dummy output for change.
-    const change = new Output();
-
-    if (this.changeAddress) {
-      change.address = this.changeAddress;
-    } else {
-      // In case we don't have a change address,
-      // we use a fake p2pkh output to gauge size.
-      change.address.fromPubkeyhash(Buffer.allocUnsafe(20));
-    }
-
-    this.tx.outputs.push(change);
-
-    // Keep recalculating the fee and funding
-    // until we reach some sort of equilibrium.
-    do {
-      const size = await this.tx.estimateSize(this.estimate);
-
-      this.fee = this.getFee(size);
-
-      if (this.maxFee > 0 && this.fee > this.maxFee)
-        throw new FundingError('Fee is too high.');
-
-      // Failed to get enough funds, add more coins.
-      if (!this.isFull())
-        this.fund();
-    } while (!this.isFull() && this.index < this.coins.length);
-  }
-
-  /**
-   * Initiate selection based on a hard fee.
-   */
-
-  selectHard() {
-    this.fee = this.hardFee;
-    this.fund();
-  }
-
-  resolveInputCoins() {
-    if (this.inputs.size === 0)
-      return;
-
-    const coins = [];
-
-    for (let i = 0 ; i < this.inputs.size; i++) {
-      coins.push(null);
-    }
-
-    // first resolve from coinview if possible.
-    for (const key of this.inputs.keys()) {
-      const prevout = Outpoint.fromKey(key);
-
-      if (this.view.hasEntry(prevout)) {
-        const coinEntry = this.view.getEntry(prevout);
-        const i = this.inputs.get(key);
-
-        if (i != null) {
-          assert(!coins[i]);
-          coins[i] = coinEntry.toCoin(prevout);
-          this.inputs.delete(key);
-        }
-      }
-    }
-
-    // Now try to resolve from the passed coins array.
-    if (this.inputs.size > 0) {
-      for (const coin of this.coins) {
-        const {hash, index} = coin;
-        const key = Outpoint.toKey(hash, index);
-        const i = this.inputs.get(key);
-
-        if (i != null) {
-          assert(!coins[i]);
-          coins[i] = coin;
-          this.inputs.delete(key);
-        }
-      }
-    }
-
-    if (this.inputs.size > 0)
-      throw new Error('Could not resolve preferred inputs.');
-
-    for (const coin of coins) {
-      this.tx.addCoin(coin);
-      this.chosen.push(coin);
-    }
-  }
-}
-
-/**
- * Default fee rate
- * for coin selection.
- * @const {Amount}
- * @default
- */
-
-CoinSelector.FEE_RATE = 10000;
-
-/**
- * Minimum fee to start with
- * during coin selection.
- * @const {Amount}
- * @default
- */
-
-CoinSelector.MIN_FEE = 10000;
-
-/**
- * Funding Error
- * An error thrown from the coin selector.
- * @ignore
- * @extends Error
- * @property {String} message - Error message.
- * @property {Amount} availableFunds
- * @property {Amount} requiredFunds
- */
-
-class FundingError extends Error {
-  /**
-   * Create a funding error.
-   * @constructor
-   * @param {String} msg
-   * @param {AmountValue} [available]
-   * @param {AmountValue} [required]
-   */
-
-  constructor(msg, available, required) {
-    super();
-
-    this.type = 'FundingError';
-    this.message = msg;
-    this.availableFunds = -1;
-    this.requiredFunds = -1;
-
-    if (available != null) {
-      this.message += ` (available=${Amount.coin(available)},`;
-      this.message += ` required=${Amount.coin(required)})`;
-      this.availableFunds = available;
-      this.requiredFunds = required;
-    }
-
-    if (Error.captureStackTrace)
-      Error.captureStackTrace(this, FundingError);
-  }
-}
-
 /*
  * Helpers
  */
 
-function sortAge(a, b) {
-  a = a.height === -1 ? 0x7fffffff : a.height;
-  b = b.height === -1 ? 0x7fffffff : b.height;
-  return a - b;
-}
-
-function sortRandom(a, b) {
-  return Math.random() > 0.5 ? 1 : -1;
-}
-
-function sortValue(a, b) {
-  if (a.height === -1 && b.height !== -1)
-    return 1;
-
-  if (a.height !== -1 && b.height === -1)
-    return -1;
-
-  return b.value - a.value;
-}
+/**
+ * @param {Input} a
+ * @param {Input} b
+ * @returns {Number}
+ */
 
 function sortInputs(a, b) {
   return a.compare(b);
 }
+
+/**
+ * @param {Output} a
+ * @param {Output} b
+ * @returns {Number}
+ */
 
 function sortOutputs(a, b) {
   return a.compare(b);
@@ -1942,9 +1468,6 @@ function sortLinked(a, b) {
  * Expose
  */
 
-exports = MTX;
-exports.MTX = MTX;
-exports.Selector = CoinSelector;
-exports.FundingError = FundingError;
+MTX.MTX = MTX;
 
-module.exports = exports;
+module.exports = MTX;
